fix: rename portfolioheader and portfoliolist
