+++ src/pages/Portfolios.tsx
@@ -1,5 +1,5 @@
 
-import { MainBody, Header, Greetings } from "../components";
+import { PortfolioList, Header, PortfolioHeader } from "../components";
 
 export const Portfolio: React.FC = () => {
     return (
@@ -7,8 +7,8 @@ export const Portfolio: React.FC = () => {
             <Header />
             <main className="w-full h-full overflow-y-auto pt-10 bg-gray-100">
                 <div className="px-12 pb-5 sm:px-24 md:px-28 lg:px-36">
-                    <Greetings />
-                    <MainBody/>
+                    <PortfolioHeader />
+                    <PortfolioList/>
                 </div>
             </main>
         </div>
