--- src/types/type.ts
@@ -1,10 +1,0 @@
-import { ButtonType, ButtonVariant } from "./enum";
-
-export type ButtonProps = {
-    label: string;
-    style?: string;
-    type?: ButtonType
-    variant?: ButtonVariant;
-    isImg:boolean;
-    onClick: () => void;
-}
